[css-grid] Remove Blink-specific code for handling orthogonal grid items https://bugs.webkit.org/show_bug.cgi?id=171807
Reviewed by Darin Adler.

Source/WebCore:

This code was added in r203252 as part of a patch improving the handling of
{min|max}-content with orthogonal flows. The original code came from Blink which performs a
pre-layout of orthogonal boxes in FrameView. That is not true in the case of WebKit so we do
not need this code at all. Actually it was causing incorrect renderings in some tests.

I'm also removing an invalid return statement just after another return.

* rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithmStrategy::minContentForChild):
(WebCore::GridTrackSizingAlgorithmStrategy::maxContentForChild):

LayoutTests:

Updated the expectations. They should now be compatible with Firefox and Chrome.

* fast/css-grid-layout/grid-track-sizing-with-orthogonal-flows.html:
* fast/css-grid-layout/grid-track-sizing-with-percentages-and-orthogonal-flows.html:
* fast/css-grid-layout/repeating-layout-must-produce-the-same-results.html:

Identifier: 188825@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216574 268f45cc-cd09-0410-ab3c-d52691b4dbfc