REGRESSION (r206960): Possible null pointer dereference under DOMSelection::getRangeAt() https://bugs.webkit.org/show_bug.cgi?id=171925 <rdar://problem/29931223>
Reviewed by Wenson Hsieh.

We have evidence that selection().firstRange() can return null in DOMSelection::getRangeAt().
When this happens, we now throw an INDEX_SIZE_ERR instead of dereferencing it.

I believe this can happen if the VisibleSelection is orphaned but not none, because
rangeCount() only checks for isNone() but VisibleSelection::firstRange() can return null
if isNoneOrOrphaned().

No new tests, I do not know how to reproduce.

* page/DOMSelection.cpp:
(WebCore::DOMSelection::getRangeAt):

Identifier: 188851@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216607 268f45cc-cd09-0410-ab3c-d52691b4dbfc