GCController.cpp's collect() should be Async https://bugs.webkit.org/show_bug.cgi?id=171708
Reviewed by Saam Barati and Geoffrey Garen.

No new tests because no change in behavior.

This is one step towards not requesting sync GCs in WebCore. I'm landing this incrementally to
make bisecting super easy.

This is a ~7% JetStream iOS "regression", because JetStream has a bug where it allows trunk
to sneakily hide GC work between when JetStream measures time. After this change, we are no
longer trying to be sneaky.

* bindings/js/GCController.cpp:
(WebCore::collect):



Identifier: 188866@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216625 268f45cc-cd09-0410-ab3c-d52691b4dbfc