[Cocoa] errors are not propagated to PassKit when calling ApplePaySession.completePayment() https://bugs.webkit.org/show_bug.cgi?id=172253 <rdar://problem/32258020>
Reviewed by Dan Bernstein.

In ApplePaySession::completePayment(), releaseReturnValue() was being called twice on the
same convertedResult. Since the first call moved the errors vector out of convertedResult,
the vector is empty in the second call. It's the second call that sends the result to the
UI process, so we end up with an empty arary when we call PassKit's delegate completion
handler.

* Modules/applepay/ApplePaySession.cpp:
(WebCore::ApplePaySession::completePayment):


Identifier: 189177@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217022 268f45cc-cd09-0410-ab3c-d52691b4dbfc